Change int for long in intavl.[hc]

+++ intavl.c
@@ -20,25 +20,25 @@ AVL_TREE new_intavl_tree()
 
 AVL_ITERATOR intavl_tree_put(
 		AVL_TREE t, /* the tree descriptor */
-		const int k, /* the key string */
+		const long k, /* the key string */
 		void *d) /* the data */
 {
 	return avl_tree_put(t, (void *)k, d);
 } /* intavl_tree_put */
 
-int intavl_tree_del(AVL_TREE t, const int k)
+int intavl_tree_del(AVL_TREE t, const long k)
 {
 	return avl_tree_del(t, (void *)k);
 } /* intavl_tree_del */
 
-int intavl_tree_has(AVL_TREE t, const int k)
+int intavl_tree_has(AVL_TREE t, const long k)
 {
 	return avl_tree_has(t, (void *)k);
 }
 
-int intavl_iterator_key(AVL_ITERATOR i)
+long intavl_iterator_key(AVL_ITERATOR i)
 {
-	return (int) avl_iterator_key(i);
+	return (long) avl_iterator_key(i);
 } /* intavl_iterator_key */
 
 /* $Id: intavl.c,v 1.1 2014/01/07 20:03:15 luis Exp $ */
